feat: Centralizar dados mock e utilitários para desenvolvimento e testes
