+++ src/__mocks__/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Dados mock centralizados para desenvolvimento e testes
+ * 
+ * Este arquivo centraliza todos os dados mock para evitar duplicação
+ * e garantir consistência entre ambiente de desenvolvimento e testes.
+ */
+
+export * from './data/user'
+export * from './data/portfolio'
+export * from './data/stocks'
+export * from './data/simulations'
+export * from './services/firebase'
+export * from './utils/helpers'
