+++ src/__tests__/calculator.test.tsx
@@ -18,8 +18,8 @@ import { TestWrapper } from './helpers/test-wrapper'
 import {
   mockAuth,
   mockAuthenticatedUser,
-  resetAllMocks
-} from './mocks/firebase'
+  resetAuthMocks
+} from '@/__mocks__'
 import {
   expectErrorToast,
   expectSuccessToast,
@@ -57,7 +57,7 @@ describe.skip('Testes da Calculadora de Balanceamento', () => {
   }
 
   beforeEach(() => {
-    resetAllMocks()
+    resetAuthMocks()
 
     // Setup usuário autenticado
     mockAuthenticatedUser()
